Setting up Animation Rigging on Player

+++ ABZero_Core/Assets/_Scripts/1_Scripts_PC/Pc_VAMT_SObj.cs
@@ -14,7 +14,7 @@ namespace ABZ_Pc
         public int      horizontalStrafePower;
         public int      horizontalBurstPower;
 
-        public int      pivotRotatioValue;
+        public float    pivotRotatioValue;
         public float    CamTurnSensitivity;
     }
 }
